Add development dependencies
Add a linter, a code formatter and a Makefile to ease development.

This was necessary as a code formatter will lead to less code-style
discussions and the linter (pylint) assures that the code is pep8
compliant and that it follows good pratices.

The use of a Makefile is mostly to help setting up the enviroment and
diminish virtual enviroment errors.